requirements.txt
lxml==4.7.1
pillow==9.0.1
protobuf==3.17.3
six==1.16.0
